enable external links in secondary windows
